% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata_documentation.R
\docType{data}
\name{MQC_testdata}
\alias{MQC_testdata}
\title{MQC_testdata}
\format{
An object of class \code{data.frame} with 26 rows and 6 columns:
\describe{
  \item{Column1}{species}
  \item{Column2}{DNA_hilo_RR}
  \item{Column3}{perfect result}
  \item{Column4}{high FPRA example}
  \item{Column5}{high diversity example}
  \item{Column6}{low sensitivity example}
}
}
\source{
Generated by the package authors for demonstration purposes.
}
\usage{
data(MQC_testdata)
}
\description{
Data used for demonstrating the functionality of the microbiomeMQC package.
}
